Added LFSR TB to the repo
- Adjusted lfsr.v so it works with verilator
   Turned off the UNOPTFLAT warnings

- Tightened down the genericfir and fastfir.v output width, to only
  what it *should* be and no extra (just to make sure it really works)

- Fixed the filtertb.cpp response function, so it now works for both
  fastfir_tb and genericfir_tb

- Updated the filtertb.cpp so that it uses long variables--this will make it
  easier to make certain the test bench doesn't overflow on reasonably sized
  filters

- Added a test involving a proper lowpass filter, to see whether or not
  the TB can handle it (it can)

- Removed the last vestigaes of the OPT_DROP_CE from delayw.v (oops)

+++ bench/cpp/boxcar_tb.cpp
@@ -100,11 +100,11 @@ public:
 		}
 	}
 
-	void	apply(int nlen, int *data) {
+	void	apply(int nlen, long *data) {
 		PARENTTB::apply(nlen, data);
 	}
 
-	void	load(int nlen, int *data) {
+	void	load(int nlen, long *data) {
 		PARENTTB::m_core->i_reset = 0;
 		PARENTTB::m_core->i_ce    = 0;
 		PARENTTB::m_core->i_tap_wr= 1;
@@ -115,7 +115,7 @@ public:
 		PARENTTB::clear_cache();
 	}
 
-	void	testload(int nlen, int *data) {
+	void	testload(int nlen, long *data) {
 		load(nlen, data);
 
 		for(int k=0; k<nlen; k++) {
